server/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
websockets==12.0
sqlalchemy==2.0.23
alembic==1.12.1
pydantic==2.5.0
bcrypt==4.0.1
python-jose==3.3.0
python-multipart==0.0.6
pydantic-settings==2.1.0
passlib==1.7.4
pymysql==1.1.0